refactor(ROADMAP §6): Phase C group-7 LOW duplication dedup (host/core)
Resolve the group-7 [LOW] duplication findings across the host/core layer,
held to docs/cleanup/bigrefactor/NAMING-STYLE-STANDARD.md. Single-source
copy-pasted scaffolds; behavior-NEUTRAL throughout (§12 parity preserved).

WHAT (one home each, no math change):
- qpadm: default_fit_precision() factory single-homes the unified emulated{40}
  fit precision (qpadm_fit.hpp); run_qpadm_impl / fit_one_model templated on the
  f2 source fold the DeviceF2Blocks/F2BlockTensor twin paths; primary_backend()
  + kPrimaryGpu name the gpus[0] index; scatter_by_pos single-homes the slot
  scatter. ranktest fuses the x_total gather + ind map; nested_models hoists the
  row-major accessor + per-column mean.
- core/internal: small_linalg cm_index / lu_solve_rhs / apply_rotation /
  col_sqnorm / col_dot single-home the column-major index, triangular solve, and
  Jacobi loops (bit-identical op order); pchisq_gamma_prefactor single-homes the
  incomplete-gamma prefactor; f2_estimator / launch_config / views / decode_af
  fold repeated widenings + a dead store.
- fstats/domain: validate_partials_scaffold hoists the shared validator skeleton;
  f2_combine place() folds the f2/vpair copy; f2_blocks_multigpu clamp_nonneg /
  nonneg_count + finish_streamed_tier; block_partition_rule idx() + a fail lambda.
- device: vram_budget nonneg() / budget_bytes() shared by tier_select; resources
  takes the visible count once; shard_plan make_shard(); host_ram tier table;
  f2_disk_format slab_offset().
- io: snp_reader parse_full<T>; ind_reader throw_io_error + filter_into;
  geno_reader hoisted streamsize + shared size-operands substring; eigenstrat
  fail() + widened packed_bytes; snp_filter require_at_least; filter_decision
  is_transition/is_transversion delegate to is_multiallelic.

GUARDRAILS: NO §3.2 protected parity vocab renamed (P/M/n_block/nl/nr/r/Q/Qinv/
A/B/b/g/s/f2/vpair untouched); POD-vs-member underscore split preserved; pointer-
wrapper + const __restrict__ buckets and the device-side kernels (qpadm_fit_kernels.cu,
block_sink.cu) NOT touched (deferred/skipped).

GATE (the box, Release -Werror clean rebuild): default ctest 39/39 green vs the
real-AADR AT2 goldens; STEPPE_THOROUGH=1 qpadm 2/2 green (CpuBackend oracle vs
golden_fit0/golden_fit1_NRBIG/golden_rot). No regression, no synthetic data.

+++ src/core/internal/views.hpp
@@ -64,7 +64,10 @@ struct MatView {
     /// reference and the GPU feeder. No bounds check (hot path); callers respect
     /// 0 ≤ i < P and 0 ≤ s < M.
     [[nodiscard]] double element(int i, long s) const noexcept {
-        return data[static_cast<long>(i) + static_cast<long>(P) * s];
+        // `P * s` already promotes to `long` (`s` is `long`), so only `i` needs the
+        // explicit widening cast; the redundant `static_cast<long>(P)` is dropped
+        // (DRY; NAMING-STYLE-STANDARD §2.5; findings group-7 7.3). Value unchanged.
+        return data[static_cast<long>(i) + P * s];
     }
 };
